requirements.txt
contourpy==1.3.1
cycler==0.12.1
editdistance==0.8.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.13.1
tzdata==2025.2
